+++ src/Phan/Language/Type.php
@@ -3107,18 +3107,21 @@ class Type
             return true;
         }
 
-        if ($type instanceof MixedType) {
-            return true;
+        $other_is_nullable = $type->isNullable();
+        // A nullable type is not a subtype of a non-nullable type
+        if ($this->isNullable() && !$other_is_nullable) {
+            return false;
         }
 
-        // A nullable type is not a subtype of a non-nullable type
-        if ($this->is_nullable && !$type->is_nullable) {
-            return false;
+        if ($type instanceof MixedType) {
+            // e.g. ?int is a subtype of mixed, but ?int is not a subtype of non-empty-mixed/non-null-mixed
+            // (check isNullable first)
+            return true;
         }
 
         // Get a non-null version of the type we're comparing
         // against.
-        if ($type->is_nullable) {
+        if ($other_is_nullable) {
             $type = $type->withIsNullable(false);
 
             // Check one more time to see if the types are equal
@@ -3248,6 +3251,7 @@ class Type
             return true;
         }
         if ($this->isNullable() && !$union_type->containsNullable()) {
+            // e.g. can't cast ?int to int, mixed to non-null-mixed, etc.
             return false;
         }
         $this_resolved = $this->withStaticResolvedInContext($context);