+++ src/Phan/Language/UnionType.php
@@ -1331,7 +1331,7 @@ class UnionType implements Serializable
         $new_type_set = [];
 
         foreach ($type_set as $type) {
-            if ($type->isNullable()) {
+            if ($type->isNullableLabeled()) {
                 $is_nullable = true;
             }
             if ($type->withIsNullable(false)->asExpandedTypes($code_base)->hasType($object_type)) {
@@ -1375,12 +1375,7 @@ class UnionType implements Serializable
     public function containsNonMixedNullable(): bool
     {
         foreach ($this->type_set as $type) {
-            if ($type->isNullable()) {
-                if ($type instanceof MixedType) {
-                    if ($type->__toString() === 'mixed') {
-                        continue;
-                    }
-                }
+            if ($type->isNullableLabeled()) {
                 return true;
             }
         }
@@ -1696,6 +1691,7 @@ class UnionType implements Serializable
      */
     public function isNonNullNumberType(): bool
     {
+        // TODO: This duplicates isNonNullIntOrFloatType
         if (\count($this->type_set) === 0) {
             return false;
         }
@@ -3615,10 +3611,28 @@ class UnionType implements Serializable
     public function numericTypes(): UnionType
     {
         // TODO: Replace mixed with int|string|float in ConditionVisitor
-        return $this->makeFromFilter(static function (Type $type): bool {
-            // IntType and LiteralStringType
-            return $type->isPossiblyNumeric();
-        });
+        return UnionType::of(
+            self::toNumericTypes($this->type_set),
+            self::toNumericTypes($this->real_type_set)
+        );
+    }
+
+    /**
+     * @param list<Type> $type_set
+     * @return list<Type> the types that can be numeric
+     */
+    private static function toNumericTypes(array $type_set): array
+    {
+        $result = [];
+        foreach ($type_set as $type) {
+            if ($type->isPossiblyNumeric()) {
+                if ($type instanceof MixedType) {
+                    return UnionType::typeSetFromString('int|string|float');
+                }
+                $result[] = $type->withIsNullable(false);
+            }
+        }
+        return $result;
     }
 
     /**
@@ -5665,7 +5679,10 @@ class UnionType implements Serializable
         return UnionType::typeSetFromString('bool');
     }
 
-    /** @return list<IntType|StringType> */
+    /**
+     * @return list<IntType|StringType>
+     * @suppress PhanPartialTypeMismatchReturn not able to infer subclass from typeSetFromString
+     */
     private static function intOrStringTypeSet(): array
     {
         static $types;
@@ -5945,7 +5962,7 @@ class UnionType implements Serializable
     public function containsDefiniteNonObjectType(): bool
     {
         foreach ($this->type_set as $type) {
-            if ($type->isNullable() || $type->isDefiniteNonObjectType()) {
+            if ($type->isNullableLabeled() || $type->isDefiniteNonObjectType()) {
                 return true;
             }
         }
@@ -5960,7 +5977,7 @@ class UnionType implements Serializable
     public function containsDefiniteNonObjectAndNonClassType(): bool
     {
         foreach ($this->type_set as $type) {
-            if ($type->isNullable() || ($type->isDefiniteNonObjectType() && !$type instanceof StringType)) {
+            if ($type->isNullableLabeled() || ($type->isDefiniteNonObjectType() && !$type instanceof StringType)) {
                 return true;
             }
         }
@@ -5973,7 +5990,7 @@ class UnionType implements Serializable
     public function containsDefiniteNonCallableType(): bool
     {
         foreach ($this->type_set as $type) {
-            if ($type->isNullable() || $type->isDefiniteNonCallableType()) {
+            if ($type->isNullableLabeled() || $type->isDefiniteNonCallableType()) {
                 return true;
             }
         }