+++ NEWS.md
@@ -9,6 +9,7 @@ New features (Analysis):
 + Raise the severity of some php 8.0 incompatibility issues to critical.
 + Fix handling of references after renaming variadic reference parameters of `fscanf`/`scanf`/`mb_convert_variables`
 + Mention if PhanUndeclaredFunction is potentially caused by the target php version being too old. (#4230)
++ Support a `non-null-mixed` type and change the way analysis involving nullability is checked for `mixed` (phpdoc and real). (#4278, #4276)
 
 Nov 27 2020, Phan 3.2.6
 -----------------------